Change price type from double to int and implemented working discounts with bagels only

+++ src/main/java/com/booleanuk/extension/Discount.java
@@ -2,9 +2,9 @@ package com.booleanuk.extension;
 
 public class Discount {
     private final int discountPackSize;
-    private final double packPriceAfterDiscount;
+    private final int packPriceAfterDiscount;
 
-    public Discount(int discountPackSize, double packPriceAfterDiscount) {
+    public Discount(int discountPackSize, int packPriceAfterDiscount) {
         this.discountPackSize = discountPackSize;
         this.packPriceAfterDiscount = packPriceAfterDiscount;
     }
@@ -13,7 +13,7 @@ public class Discount {
         return discountPackSize;
     }
 
-    public double getPackPriceAfterDiscount() {
+    public int getPackPriceAfterDiscount() {
         return packPriceAfterDiscount;
     }
 }
